docker/test/requirements.txt
urllib3
requests
HTMLTestRunner-Python3